Adding ability to map from OS indexes to PU or Core indexes based on hwloc data. This is necessary because Fugaku OS reports that the cores available are 12-59, but the Cores are actually indexed 0-47.

+++ examples/main.cpp
@@ -74,7 +74,7 @@ int main(int argc, char *argv[]){
         iterations = atol(argv[1]);
     }
 
-    for (int iter = 0 ; iter < iterations ; iter++) {
+    for (size_t iter = 0 ; iter < iterations ; iter++) {
     /* do some work */
 #ifdef NDEBUG
         constexpr int n = 1024*4;
